requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
django-markup==1.2
django-unslashed==0.3.0
gitdb2==2.0.2
GitPython==2.1.5
idna==2.5
inflection==0.3.1
MarkupSafe==1.0
mediawiki-parser==0.4.1
networkx==1.11
numpy==1.13.1
pandas==0.20.3
pijnu==20160727
psycopg2==2.7.3
py3compat==0.4
Pygments==2.2.0
python-creole==1.3.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.3
six==1.10.0
smmap2==2.0.3
urllib3==1.22
django-elasticsearch-dsl
elasticsearch-dsl>=5.0,<6.0
